Move decision making functions to new module
